+++ packages/general/src/util/Mutex.ts
@@ -5,6 +5,7 @@
  */
 
 import { Logger } from "../log/Logger.js";
+import { asError } from "./Error.js";
 
 const logger = Logger.get("Mutex");
 
@@ -30,8 +31,8 @@ export class Mutex implements PromiseLike<unknown> {
      * await.
      */
     then<TResult1 = void, TResult2 = never>(
-        onfulfilled?: ((value: unknown) => TResult1 | PromiseLike<TResult1>) | undefined | null,
-        onrejected?: ((reason: any) => TResult2 | PromiseLike<TResult2>) | undefined | null,
+        onfulfilled?: ((value: unknown) => TResult1 | PromiseLike<TResult1>) | null,
+        onrejected?: ((reason: any) => TResult2 | PromiseLike<TResult2>) | null,
     ): PromiseLike<TResult1 | TResult2> {
         return (this.#promise || Promise.resolve()).then(onfulfilled, onrejected);
     }
@@ -74,7 +75,7 @@ export class Mutex implements PromiseLike<unknown> {
                 try {
                     resolve(await task());
                 } catch (e) {
-                    reject(e);
+                    reject(asError(e));
                 }
             }, cancel);
         });